refactor: remove unused tests and constants from test_sqlalchemy.py
